+++ cmd/sfreleaser/release.go
@@ -56,7 +56,7 @@ var ReleaseCmd = Command(release,
 		flags.StringArray("pre-build-hooks", nil, "Set of pre build hooks to run before run the actual building steps")
 		flags.String("upload-substreams-spkg", "", "If provided, add this Substreams package file to the release, if manifest is a 'substreams.yaml' file, the package is first built")
 		flags.Bool("publish-now", false, "By default, publish the release to GitHub in draft mode, if the flag is used, the release is published as latest")
-		flags.String("goreleaser-docker-image", "goreleaser/goreleaser-cross:v1.20.3", "Full Docker image used to run Goreleaser tool (which perform Go builds and GitHub releases (in all languages))")
+		flags.String("goreleaser-docker-image", "goreleaser/goreleaser-cross:v1.20.5", "Full Docker image used to run Goreleaser tool (which perform Go builds and GitHub releases (in all languages))")
 
 		// Brew Flags
 		flags.Bool("brew-disabled", false, "[Brew only] Disable Brew tap release completely, only applies for 'Golang'/'Application' types")
@@ -132,7 +132,7 @@ func release(cmd *cobra.Command, args []string) error {
 	verifyTools()
 
 	if release.Version == "" {
-		release.Version = promptVersion()
+		release.Version = promptVersion(readReleaseNotesVersion(changelogPath))
 	}
 
 	// For simplicity in the code below
@@ -166,7 +166,7 @@ func release(cmd *cobra.Command, args []string) error {
 		executeHooks(preBuildHooks, global, release)
 	}
 
-	uploadSpkgPath := prepareSubstreamsSpkg(uploadSubstreamsSPKG, global, release)
+	uploadSpkgPath := prepareSubstreamsSpkg(uploadSubstreamsSPKG, global, release.Version)
 
 	fmt.Println()
 	fmt.Println("Creating temporary tag so that goreleaser can work properly")
@@ -260,13 +260,13 @@ func executeHook(command string, model map[string]any) {
 	run(out.String())
 }
 
-func prepareSubstreamsSpkg(spkgPath string, global *GlobalModel, release *ReleaseModel) string {
+func prepareSubstreamsSpkg(spkgPath string, global *GlobalModel, version string) string {
 	if spkgPath == "" {
 		return ""
 	}
 
 	if !strings.HasSuffix(spkgPath, ".spkg") {
-		spkgPath = filepath.Join(global.WorkingDirectory, "build", global.Project+"-"+release.Version+".spkg")
+		spkgPath = filepath.Join(global.WorkingDirectory, "build", global.Project+"-"+version+".spkg")
 
 		fmt.Printf("Packing your Substreams file at %q\n", spkgPath)
 		run("substreams pack -o", "'"+spkgPath+"'")