+++ cmd/sfreleaser/release_notes.go
@@ -10,7 +10,54 @@ import (
 	"go.uber.org/zap"
 )
 
-var headerRegex = regexp.MustCompile(`^##([^#])`)
+var headerRegex = regexp.MustCompile(`^##([^#]+)`)
+
+func readReleaseNotesVersion(changelogFile string) string {
+	if !cli.FileExists(changelogFile) {
+		return ""
+	}
+
+	file, err := os.Open(changelogFile)
+	cli.NoError(err, "Unable to open changelog %q", changelogFile)
+	defer file.Close()
+
+	scanner := bufio.NewScanner(file)
+	for scanner.Scan() {
+		line := scanner.Text()
+		if headerRegex.MatchString(line) {
+			return extractVersionFromHeader(line)
+		}
+	}
+
+	cli.NoError(scanner.Err(), "Unable to scan lines from changelog")
+
+	// We found nothing!
+	return ""
+}
+
+func extractVersionFromHeader(header string) string {
+	matches := headerRegex.FindAllStringSubmatch(header, 1)
+	if len(matches) != 1 {
+		return ""
+	}
+
+	groups := matches[0]
+	if len(groups) != 2 {
+		return ""
+	}
+
+	version := strings.TrimSpace(groups[1])
+	if version == "" {
+		return ""
+	}
+
+	normalizedVersion := strings.ToLower(version)
+	if normalizedVersion != "unreleased" && normalizedVersion != "next" {
+		return version
+	}
+
+	return ""
+}
 
 func readReleaseNotes(changelogFile string) string {
 	if !cli.FileExists(changelogFile) {